feat: generalize Chip for both flags and core classes
