+++ src/views/components/injected/CourseCatalogInjectedPopup/HeadingAndActions.tsx
@@ -3,8 +3,8 @@ import type { Course } from '@shared/types/Course';
 import type Instructor from '@shared/types/Instructor';
 import type { UserSchedule } from '@shared/types/UserSchedule';
 import { Button } from '@views/components/common/Button';
-import { Chip, flagMap } from '@views/components/common/Chip';
-import { CoreChip, coreMap } from '@views/components/common/CoreChip';
+import { Chip, coreMap, flagMap } from '@views/components/common/Chip';
+// import { CoreChip, coreMap } from '@views/components/common/CoreChip';
 import Divider from '@views/components/common/Divider';
 import Link from '@views/components/common/Link';
 import Text from '@views/components/common/Text/Text';
@@ -153,12 +153,16 @@ export default function HeadingAndActions({ course, activeSchedule, onClose }: H
                             <Chip
                                 key={flagMap[flag as keyof typeof flagMap]}
                                 label={flagMap[flag as keyof typeof flagMap]}
+                                variant='flag'
+                                labelMap={flagMap}
                             />
                         ))}
                         {core.map((coreVal: string) => (
-                            <CoreChip
+                            <Chip
                                 key={coreMap[coreVal as keyof typeof coreMap]}
                                 label={coreMap[coreVal as keyof typeof coreMap]}
+                                variant='core'
+                                labelMap={coreMap}
                             />
                         ))}
                     </div>
